refactor(mapmory): 지도 요약 패키지 책임별 분리
