+++ backend/src/main/java/com/mapmory/backend/member/repository/MemberRepository.java
@@ -1,5 +1,6 @@
-package com.mapmory.backend.member;
+package com.mapmory.backend.member.repository;
 
+import com.mapmory.backend.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
