+++ backend/src/main/java/com/mapmory/backend/travelrecord/mapsummary/policy/MapColorLevel.java
@@ -1,4 +1,4 @@
-package com.mapmory.backend.travelrecord.mapsummary;
+package com.mapmory.backend.travelrecord.mapsummary.policy;
 
 public enum MapColorLevel {
     NONE,
